+++ CHANGELOG.md
@@ -8,9 +8,13 @@ Releases prior to January 2023 are tracked on the project GitHub [Releases Page]
 
 ### Changed
 
-- Use proper namespace to allow declaring content modules in different namespaces in the same plugin ([IJPL-245093](https://youtrack.jetbrains.com/issue/IJPL-245093))
+- Migrate JAXB from `javax` to `jakarta` namespace. ([#1486](https://github.com/JetBrains/intellij-plugin-verifier/pull/1486))
 - `PropertyWithDefaultValue` class is now `com.jetbrains.plugin.structure.base.problems.PropertyWithDefaultValue`
 - `DescriptionNotStartingWithLatinCharacters` class is now `DescriptionNotStartingWithLatinCharacters`
+- Move description validator to the base validator ([#1494](https://github.com/JetBrains/intellij-plugin-verifier/pull/1494))
+- Use proper namespace to allow declaring content modules in different namespaces in the same plugin ([IJPL-245093](https://youtrack.jetbrains.com/issue/IJPL-245093))
+- Update ASM to 9.10
+- Update Gradle to 9.5.1
 
 ### Fixed